rename version to blob_storage
