+++ src/blob_storage.h
